Seperate logic into several files. Create "services" dir for applied classes. Use abs path in db_image_uri_fetcher (func as main), and relative paths in module interactions.
